move repointerface to contract
