+++ src/Controller/GroupController.php
@@ -5,8 +5,8 @@ namespace Xsanisty\UserManager\Controller;
 use Exception;
 use Xsanisty\Admin\DashboardModule;
 use Cartalyst\Sentry\Users\UserInterface;
-use Xsanisty\UserManager\Repository\GroupRepositoryInterface;
-use Xsanisty\UserManager\Repository\PermissionRepositoryInterface;
+use Xsanisty\UserManager\Contract\GroupRepositoryInterface;
+use Xsanisty\UserManager\Contract\PermissionRepositoryInterface;
 
 class GroupController
 {
